move counter to be nested
