+++ convex/convex.config.ts
@@ -1,5 +1,5 @@
 import { defineApp } from "convex/server";
-import counter from "../counter/component/convex.config";
+import counter from "./counter/component/convex.config.js";
 
 const app = defineApp();
 app.use(counter);
